ContentViewのリファクタリング
- sendInfoを削除してContentView内でprivate funcにした(sendInfo関数をContentView以外で使わないため)
